refactor: rename eventstore references to messagestore
