+++ services/ingestor/src/main/resources/application.conf
@@ -9,7 +9,7 @@ ktor {
     }
 }
 
-event-store {
+messagestore {
     producer {
         "bootstrap.servers" = "localhost:9092"
         "bootstrap.servers" = ${?KAFKA_PC_BOOTSTRAP_SERVERS}
