+++ shared/messagestore/gradle.properties
