+++ shared/messagestore/src/main/resources/logback.xml
