+++ shared/messagestore/src/test/kotlin/company/thewebhook/eventstore/ApplicationTest.kt
@@ -1,4 +1,4 @@
-package company.thewebhook.eventstore
+package company.thewebhook.messagestore
 
 import kotlin.test.*
